twitoff/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
basilica==0.2.6
beautifulsoup4==4.8.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.2
Pillow==6.1.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
PySocks==1.7.0
python-decouple==3.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.8
tweepy==3.8.0
urllib3==1.24.2
Werkzeug==0.15.5
wrapt==1.11.2
